refactor(poll): use navigator
