+++ package.json
@@ -77,6 +77,7 @@
     "@babel/core": "^7.18.6",
     "@babel/eslint-parser": "^7.18.9",
     "@babel/eslint-plugin": "^7.18.10",
+    "@babel/plugin-syntax-flow": "^7.18.6",
     "@babel/preset-flow": "^7.18.6",
     "eslint": "^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
