+++ src/screens/poll-screen/create-poll-screen/receiving-answers-screen/styles.js
