--- src/screens/poll-screen/create-poll/styles.js
@@ -1,57 +1,0 @@
-import styled from 'styled-components/native';
-import button from '../../../components/button';
-import textInput from '../../../components/text-input';
-import Colors from '../../../constants/colors';
-
-const ButtonsContainer = styled.View``;
-
-const OptionsButton = styled(button)`
-  background-color: ${Colors.lightGray200}
-  color: ${Colors.lightGray400};
-  font-size: 16px;
-  font-weight: 400;
-  padding: 12px;
-  border-radius: 12px;
-
-  ${({ selected }) => selected
-    && `
-      background-color: #003399;
-      color: ${Colors.white};
-  `}
-`;
-
-const ConfirmButton = styled(button)`
-  background-color: ${Colors.orange};
-  color: ${Colors.white};
-  font-size: 16px;
-  font-weight: 400;
-  padding: 12px;
-  margin-bottom: 16px;
-  border-radius: 12px;
-  margin-top: 32px;
-`;
-
-const Title = styled.Text`
-  font-size: 24px;
-  font-weight: 600;
-  text-align: center;
-  padding-bottom: 24px;
-`;
-
-const AnswerTitle = styled.Text`
-  font-weight: 500;
-  font-size: 18px;
-  padding: 24px 0;
-  text-align: center;
-`;
-
-const TextInput = styled(textInput)``;
-
-export default {
-  Title,
-  OptionsButton,
-  AnswerTitle,
-  ButtonsContainer,
-  ConfirmButton,
-  TextInput,
-};
