+++ src/screens/poll-screen/navigator.js
@@ -1,11 +1,37 @@
-import React from 'react';
+import { useEffect } from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import CreatePollView from './create-poll';
-import ReceivingAnswers from './create-poll/receiving-answers';
-import PollScreen from './index';
+import { useNavigation } from '@react-navigation/native';
+import { useSelector } from 'react-redux';
+import { selectActivePoll } from '../../store/redux/slices/polls';
+import CreatePollScreen from './create-poll-screen';
+import ReceivingAnswersScreen from './create-poll-screen/receiving-answers-screen';
+import PreviousPollsScreen from './previous-polls-screen';
+import AnswerPollScreen from './answer-poll-screen';
+import { isPresenter } from '../../store/redux/slices/current-user';
+import { hasCurrentPollSelector } from '../../store/redux/slices/current-poll';
 
 const PollNavigator = () => {
   const Stack = createStackNavigator();
+  const activePollObject = useSelector(selectActivePoll);
+  const hasCurrentPoll = useSelector(hasCurrentPollSelector);
+  const amIPresenter = useSelector(isPresenter);
+  const navigation = useNavigation();
+
+  useEffect(() => {
+    if (!activePollObject && amIPresenter && hasCurrentPoll) {
+      navigation.replace('ReceivingAnswersScreen');
+    }
+    else if (!activePollObject && amIPresenter && !hasCurrentPoll) {
+      navigation.replace('CreatePollScreen');
+    }
+    else if (!activePollObject && !amIPresenter && !hasCurrentPoll) {
+      navigation.replace('PreviousPollsScreen');
+    }
+    else if (activePollObject && !amIPresenter) {
+      navigation.replace('AnswerPollScreen');
+    }
+  }, [activePollObject, amIPresenter, hasCurrentPoll]);
+
   return (
     <Stack.Navigator
       screenOptions={{
@@ -15,9 +41,10 @@ const PollNavigator = () => {
         }
       }}
     >
-      <Stack.Screen name="PollInitialScreen" component={PollScreen} />
-      <Stack.Screen name="CreatePollScreen" component={CreatePollView} />
-      <Stack.Screen name="ReceivingAnswersScreen" component={ReceivingAnswers} />
+      <Stack.Screen name="CreatePollScreen" component={CreatePollScreen} />
+      <Stack.Screen name="AnswerPollScreen" component={AnswerPollScreen} />
+      <Stack.Screen name="ReceivingAnswersScreen" component={ReceivingAnswersScreen} />
+      <Stack.Screen name="PreviousPollsScreen" component={PreviousPollsScreen} />
     </Stack.Navigator>
   );
 };
