+++ src/screens/poll-screen/previous-polls-screen/poll-card/index.js
@@ -1,13 +1,19 @@
 import { View } from 'react-native';
+import { useTranslation } from 'react-i18next';
 import Styled from './styles';
 
 const PreviousPollCard = (props) => {
   const { pollObj } = props;
   const timestamp = new Date(parseInt(pollObj.id.split('/')[2], 10));
+  const { t } = useTranslation();
 
   return (
     <Styled.ContainerPollCard>
-      <Styled.QuestionText>{pollObj.questionText}</Styled.QuestionText>
+      <Styled.QuestionText>
+        {pollObj.questionText === ''
+          ? t('mobileSdk.poll.noQuestionTextProvided')
+          : pollObj.questionText}
+      </Styled.QuestionText>
       <Styled.TimestampText>
         {`${String(timestamp.getHours()).padStart(2, '0')}:${String(
           timestamp.getMinutes()
@@ -16,9 +22,17 @@ const PreviousPollCard = (props) => {
       {pollObj.answers.map((answer) => (
         <View key={answer.id}>
           <Styled.AnswerContainer>
-            <Styled.KeyText>{answer.key} : </Styled.KeyText>
             <Styled.KeyText>
-              {(pollObj.numResponders == 0 ? 0 : (answer.numVotes / pollObj.numResponders) * 100).toFixed(0)}%
+              {answer.key}
+              {' '}
+              :
+              {' '}
+            </Styled.KeyText>
+            <Styled.KeyText>
+              {(pollObj.numResponders === 0
+                ? 0
+                : (answer.numVotes / pollObj.numResponders) * 100).toFixed(0)}
+              %
             </Styled.KeyText>
           </Styled.AnswerContainer>
         </View>
